Set maximum SPI speed for MCP23S17
The linux-raspberrypi kernel update to 4.9.43 included a change to the
default reported max speed of the linux spidev, from 100KHz to 125MHz.

kernel: BCM270X_DT: Set spidev spi-max-frequency to 125MHz.
See https://github.com/raspberrypi/linux/issues/2165.

It appears the reported max speed is used as the default if you don't
specify it for the device. The MCP23S17 chip used in the PiFace digital
board can only handle 10MHz but we were using the default so it was not
working after the recent kernel update. This change explicitly sets the
maximum speed after opening the device.